// api/src/core/models/Address/Address.d.ts
export declare class Address {
    readonly id: string;
    readonly country: string;
    readonly city: string;
    readonly street: string;
    readonly houseNumber: string;
    readonly numberOfApartment: number | null;
    readonly postcode: number;
    constructor(id: string, country: string, city: string, street: string, houseNumber: string, numberOfApartment: number | null, postcode: number);
}
//# sourceMappingURL=Address.d.ts.map